Reformated file structure.
