--- src/main/java/Task.java
@@ -1,2 +1,0 @@
-public class Task {
-}
