+++ src/main/java/app/Task.java
@@ -1,0 +1,4 @@
+package app;
+
+public interface Task {
+}
