+++ src/main/java/app/TaskList.java
@@ -1,0 +1,9 @@
+package app;
+
+import java.util.List;
+import java.util.ArrayList;
+
+public class TaskList {
+    private final List<String> taskList= new ArrayList<String>();
+
+}
