+++ src/main/java/utilities/Pair.java
@@ -1,3 +1,5 @@
+package utilities;
+
 public class Pair<T, U>{
     private final T first;
     private final U second;
